requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
joblib==1.5.1
numpy==2.0.2
scikit-learn==1.6.1
psycopg2-binary
gunicorn
python-decouple